Fixed some things

+++ main.py
@@ -2,8 +2,8 @@
 import discord
 import os
 import keep_alive
-import asyncio
 import typing
+import asyncio
 from discord.ext import commands
 from discord.ext.commands import errors
 from pretty_help import PrettyHelp
@@ -17,11 +17,19 @@ intents.members = True
 
 bot = commands.Bot(command_prefix='$', help_command=PrettyHelp(), description=description, intents=intents)
 
+# bots
+music_bot = None
 
-async def send_warning(message, title, description):
-  formatted_description = (
-    ':no_entry_sign:   **{0}**\n\n{1}'.format(title, description))
-  warning = discord.Embed(title='Warning!', description=formatted_description, color = 0xff0000)
+# channels
+ch_music = None
+ch_log = None
+ch_commands = None
+ch_bot_test = None
+
+
+async def send_warning(message, warning_type, title, description):
+  formatted_description = ':no_entry_sign:   **{0}**\n\n{1}'.format(title, description)
+  warning = discord.Embed(title=warning_type, description=formatted_description, color = 0xff0000)
 
   await message.reply(embed=warning, delete_after=15)
   await asyncio.sleep(15)
@@ -37,12 +45,40 @@ async def send_command_usage(ctx):
   
   embed = discord.Embed(title=ctx.command.name)
   embed.add_field(name='Usage', value=usage, inline=True)
-  await ctx.send(embed=embed)
+  await ctx.message.delete()
+  await ctx.send(embed=embed, delete_after=20)
+
+
+def extract_args(ctx):
+  args = ''
+  for arg in ctx.args:
+    if not arg is ctx.args[0]:
+      if isinstance(arg, discord.TextChannel):
+        args += f'#{arg} '
+      elif isinstance(arg, discord.User):
+        args += f'@{arg} '
+      elif isinstance(arg, str):
+        args += f'\'{arg}\' '
+      else:
+        args += f'{arg} '
+  return args
 
 
 @bot.event
 async def on_ready():
   print('Logged in as', bot.user)
+
+  global music_bot
+  global ch_music
+  global ch_log
+  global ch_commands
+  global ch_bot_test
+
+  music_bot = bot.get_user(235088799074484224)
+  ch_music = bot.get_channel(821946547767345152)
+  ch_log = bot.get_channel(827374735830286347)
+  ch_commands = bot.get_channel(827374609234001970)
+  ch_bot_test = bot.get_channel(827374679928471592)
 
 
 @bot.event
@@ -50,43 +86,40 @@ async def on_command_error(ctx, exception):
   if isinstance(exception, errors.CommandNotFound):
     title = 'Unknown command!'
     description = 'The given command could not be found.'
-    await send_warning(ctx.message, title, description)
+    await send_warning(ctx.message, 'Notice!', title, description)
   elif isinstance(exception, discord.ext.commands.BadArgument):
     await send_command_usage(ctx)
     
 
 @bot.listen()
 async def on_message(message):
-  if not message.channel.name == 'music':
+  if not message.channel is ch_music:
     if message.content.startswith(tuple(music_commands.commands)):
-      music_bot = bot.get_user(235088799074484224)
-      ch_music = bot.get_channel(821946547767345152)
       title = 'Wrong channel!'
       description = 'You can only use {0.mention}\'s command in {1.mention}.'.format(music_bot, ch_music)
-      
-      await send_warning(message, title, description)
+
+      await send_warning(message, 'Warning!', title, description)
 
 
 @bot.listen()
 async def on_command(ctx):
-  ch_log = bot.get_channel(827374735830286347)
   user = ctx.author
   command = ctx.command
   channel = ctx.channel
-  log_msg = '{0.mention} used `{1}` in {2.mention}'.format(user, command, channel)
-  await ch_log.send(log_msg, allowed_mentions=discord.AllowedMentions.none())
+  log_msg = f'{user.mention} used in {channel.mention}```${command} {extract_args(ctx).strip()}```'
+  embed = discord.Embed(description=log_msg)
+
+  await ch_log.send(embed=embed, allowed_mentions=discord.AllowedMentions.none())
 
 
 @bot.check
 async def pre_check(ctx):
-  if ctx.channel.name == 'commands' or ctx.channel.name == 'bot-test':
+  if ctx.channel is ch_commands or ctx.channel is ch_bot_test:
     return True
-  ch_commands = bot.get_channel(827374609234001970)
-  ch_bot_test = bot.get_channel(827374679928471592)
   title = 'Wrong channel!'
   description = 'You can only use commands in {0.mention} and {1.mention}.'.format(ch_commands, ch_bot_test)
   
-  await send_warning(ctx.message, title, description)
+  await send_warning(ctx.message, 'Warning!', title, description)
 
 
 _ping_brief='For dumbasses'
@@ -130,7 +163,7 @@ async def _pin(ctx):
   if last_user_made_message:
     await last_user_made_message.pin()
   else:
-    await send_warning(ctx.message, 'Couldn\'t find any pinnable messages in the last 10 entries.')
+    await send_warning(ctx.message, 'Notice!', '', 'Couldn\'t find any pinnable messages in the last 10 entries.')
 
 
 _clear_brief='Erases messages'
@@ -138,18 +171,23 @@ _clear_desc='Deletes past messages based on given number.'
 @bot.command(name='clear', aliases=['erase', 'clean'], brief=_clear_brief, description=_clear_desc)
 async def _clear(ctx, channel : typing.Optional[discord.TextChannel], count : int = 1):
   await ctx.message.delete()
-
+  
   if not channel:
     channel = ctx.channel
 
-  history=channel.history(limit=count)
-  async for message in history:
-    await message.delete()
+  if channel is ch_log:
+    count += 1
 
+  i = 0
+  history=channel.history(limit=count, oldest_first=False)
+  async for message in history:
+    if channel is ch_log and i == 0:
+      i += 1
+    else:
+      await message.delete()
 
 if __name__ == '__main__':
   # keeps the bot running
   keep_alive.keep_alive()
-
   # starts the bot
   bot.run(os.getenv('TOKEN'))
